Add files via upload
The design was implemented using ModelSim for simulation and synthesised in Quartus Prime. The design uses just an always_ff(Sequential)  block to implement the multiplier. The design used 13 states enumerated namely:  1. Idle: This state is achieved when a reset is triggered (The reset is a negedge triggered reset). Ready is asserted high ‘1’ that gets triggered during the posedge of start. 2. Start: In this state, the system is asserting ready to ‘0’ and the state goes to loada. The inputs appear in the next clock cycle as the output ready triggers the inputs to appear at ‘a’ in the subsequent clock cycles. 3. Loada: In this state, the system expects the first input to appear at input ‘a’. The sign bit, signed exponent and mantissas are extracted and stored in logic sign_bit_a(31st bit of ‘a’) , sign_exponent_a(30th to 23rd bit of ‘a’) and mantissa_a(22nd to 0th bit of ‘a’) . The mantissa_a takes the value of mantissa ie 22nd to 0th bit along with a binary 1 at the MSB. 4. Loadb: In this state, the system expects the second input to appear at input ‘a’. The sign bit, signed exponent and mantissas are extracted and stored in logic sign_bit_b(31stbit of ‘a’) , sign_exponent_b(30th to 23rd bit of ‘a’) and mantissa_b(22nd to 0th bit of ‘a’) . The mantissa_b takes the value of mantissa ie 22nd to 0th bit along with a binary 1 at the MSB. 5. Waitin: In this state, logic X and logic Y a pair of 24 bit registers takes the value of mantissa_a and mantissa_b . 6. Multip: In this state, the values stored in logic X and logic Y get multiplied and the resulting product is stored in a 48-bit register multip_P.  7. Waitin2: This is a redundant state that waits for a clock cycle. This state was used to get an idea about the ongoing multiplication operation during simulation. 8. Sign_comp: This state compares the sign bit of the inputs , does an XOR operation on sign_bit_a and sign_bit_b .If the subsequent result is ‘1’ ,binary 1 is stored in logic sign_bit_p else binary 0 is stored in sign_bit_p . 9. Exponent_add: This state adds the signed exponent bits of input and stores it in  sign_expoonent_p. 
210. Normalisation: In this state, if the 47th bit of multip_p is ‘0’, it shifts the bits of multip_P one bit to the right and increments the sign_exponent_p by 1.  11. Rounding: In this state, the bits of multip_p from 45th to 23rd bits are stored in 22-bit logic mantissa_p truncating rest of the bits. 12. Removebias: In this state, the bias of 127 is removed from the bits in logic sign_exponent_p. 13. Stopped: In this state the bits of sign_bit_p, sign_exponent_p and mantissa_p and concatenated and stored in output product.  A special case is added for zero case to give product 0 and another special case to display “NaN’ if the product is Not a number (sign_exponent_p is all ‘1’). Ready was also asserted to’1’ The design was purposefully made as a sequential circuit as there are no operations that would benefit the usage of a combinational block. It created the possibility of latches and other unwanted problems during Quartus synthesis.